Adds new test directory and tests, WIP
